+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/RadioactiveGeneratorScreen.java
@@ -32,8 +32,6 @@ public class RadioactiveGeneratorScreen extends AbstractContainerScreen<Radioact
 
         titleLabelX = (180 - Minecraft.getInstance().font.width(title.getString())) / 2;
         titleLabelY = 2;
-
-        inventoryLabelY = imageHeight - 92;
     }
 
     @Override