+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/RocketScreen.java
@@ -31,8 +31,6 @@ public class RocketScreen extends AbstractContainerScreen<RocketMenu> {
 
         titleLabelX = (180 - Minecraft.getInstance().font.width(title.getString())) / 2;
         titleLabelY = 2;
-
-        inventoryLabelY = imageHeight - 92;
     }
 
     @Override