+++ changelog.md
@@ -4,6 +4,12 @@
 - Add diesel
 - Add diesel generator
 - Add new config system
+- Add power banks
+- Add fluid tanks
+- Add hydrogen and oxygen to tag
+- Add solerium
+- Add corronium
+- Add tharsite
 
 ## Changes
 - Rover now need diesel instead of fuel